テストの失敗を修正: データカバレッジリポジトリのスキーマとSQL型の問題を解決
## 修正内容

### 1. テストスキーマをデータベーススキーマに合わせて修正
- `conversations` テーブルのカラム名を修正:
  - `meeting_id` → `minutes_id`
  - `content` → `comment`
  - `sequence` → `sequence_number`
- 欠落していたカラムを追加 (chapter_number, sub_chapter_number, created_at, updated_at)

### 2. テストデータのINSERT文を修正
- カラム名の変更に対応
- `test_get_speaker_matching_stats_with_data` に欠けていた minutes レコードを追加

### 3. SQLiteでの浮動小数点演算の問題を修正
- `CAST(... AS NUMERIC)` → `CAST(... AS REAL)` に変更
- 影響範囲:
  - `get_governing_body_stats()` の coverage_percentage 計算
  - `get_meeting_stats()` の avg_conversations 計算
  - `get_speaker_matching_stats()` の matching_rate, linkage_rate 計算

## 解決したテスト失敗
- test_get_governing_body_stats_with_data
- test_get_meeting_stats_empty
- test_get_meeting_stats_with_data
- test_get_meeting_stats_conference_without_meetings
- test_get_speaker_matching_stats_with_data

+++ src/infrastructure/persistence/data_coverage_repository_impl.py
@@ -100,7 +100,7 @@ class DataCoverageRepositoryImpl(IDataCoverageRepository):
                 with_meetings,
                 CASE
                     WHEN total > 0
-                    THEN ROUND(CAST(with_meetings AS NUMERIC) / total * 100, 2)
+                    THEN ROUND(CAST(with_meetings AS REAL) / total * 100, 2)
                     ELSE 0.0
                 END as coverage_percentage
             FROM stats
@@ -139,7 +139,7 @@ class DataCoverageRepositoryImpl(IDataCoverageRepository):
                 CASE
                     WHEN COUNT(DISTINCT m.id) > 0
                     THEN ROUND(
-                        CAST(COUNT(c.id) AS NUMERIC) / COUNT(DISTINCT m.id), 2
+                        CAST(COUNT(c.id) AS REAL) / COUNT(DISTINCT m.id), 2
                     )
                     ELSE 0.0
                 END as avg_conversations
@@ -217,7 +217,7 @@ class DataCoverageRepositoryImpl(IDataCoverageRepository):
                 CASE
                     WHEN ss.total_speakers > 0
                     THEN ROUND(
-                        CAST(ss.matched_speakers AS NUMERIC)
+                        CAST(ss.matched_speakers AS REAL)
                         / ss.total_speakers * 100, 2
                     )
                     ELSE 0.0
@@ -227,7 +227,7 @@ class DataCoverageRepositoryImpl(IDataCoverageRepository):
                 CASE
                     WHEN cs.total_conversations > 0
                     THEN ROUND(
-                        CAST(cs.linked_conversations AS NUMERIC)
+                        CAST(cs.linked_conversations AS REAL)
                         / cs.total_conversations * 100, 2
                     )
                     ELSE 0.0
